contract

+++ src/main/java/com/yimei/finance/representation/admin/group/GroupObject.java
@@ -20,7 +20,7 @@ public class GroupObject implements Serializable {
     @NotBlank(message = "组名不能为空", groups = {CreateGroup.class})
     private String name;
     private String type;
-    private Integer memberNums;
+    private Long memberNums;
 
 
 }
